weather_notify/requirements.txt
# 核心依赖
python-dotenv>=1.0.0  # 用于管理环境变量
requests>=2.31.0  # 用于发送HTTP请求获取天气数据
APScheduler>=3.10.1  # 用于管理定时任务
pytz>=2023.3  # 用于处理时区

# 邮件发送
smtplib3>=0.0.1  # SMTP客户端库
email>=4.0.2  # 用于构建邮件消息

# 日志管理
loguru>=0.7.0  # 更好的日志记录工具

# 开发工具
pytest>=7.4.0  # 用于单元测试
black>=23.7.0  # 代码格式化工具
flake8>=6.1.0  # 代码检查工具